+++ examples/finder_from_conf/ex_test5_yml.yml
@@ -1,3 +1,3 @@
 Keys:
-  - TestStruct2:
+  - FinderTestStruct2:
     - NonExist
