+++ examples/finder_from_conf/ex_test_invalid_yml.yml
@@ -2,8 +2,8 @@ Keys:
   - Int64
   - Float64
   - String
-  - TestStruct2Ptr:
+  - FinderTestStruct2Ptr:
     - String:
-    - TestStruct3:
+    - FinderTestStruct3:
       - String
       - Int
